Cleared an issue where uploading a new file, doesn't clear the first uploaded pdf
Automatically re-extract if a new file is uploaded

A smarter route is to track the filename. Store the filename or hash in session, and if a new one is uploaded, reprocess:

Add this before extracting:

current_filename = uploaded_pdf.name

if st.session_state.get("pdf_filename") != current_filename:
    st.session_state.pop("pdf_text", None)  # Remove old content
    st.session_state["pdf_filename"] = current_filename

That way, every time a new file is uploaded, text gets re-extracted.